Various security, code quality fixes from Covertiy scan. (#2881)
* Various security, code quality fixes from Covertiy scan.

* Brakeman rake tasks, rescan w/ ignore.

* Use bundler to execute rake.

+++ app/models/company.rb
@@ -547,7 +547,7 @@ class Company < ActiveRecord::Base
     return nil if id.nil? && code.blank?
 
     if code.blank?
-      id.destroy 
+      id.destroy unless id.nil?
       self.system_identifiers.reload
       return nil
     else
